2751.cpp

+++ 2557.c/4153.c
@@ -2,11 +2,11 @@
 
 int main()
 {
-    long a, b, c;
+     int a, b, c;
 
     while (1)
     {
-        scanf("%ld %ld %ld", &a, &b, &c);
+        scanf("%d %d %d", &a, &b, &c);
         if (a == 0 && b == 0 && c == 0) //마지막줄
             break;
